Move lite parsing into its own file
